Revert "[QA-543] JWT signing in SPOT script"
