+++ .github/workflows/pre-merge-checks.yml
@@ -6,9 +6,6 @@ on:
   merge_group:
   workflow_dispatch:
 
-permissions:
-  packages: read
-
 env:
   DOCKER_PATH: ./deploy
   TEST_TAG: user/app:test
@@ -33,9 +30,6 @@ jobs:
         uses: actions/setup-node@cdca7365b2dadb8aad0a33bc7601856ffabcc48e # v4.3.0
         with:
           node-version: 20
-          cache: npm
-          cache-dependency-path: "deploy/scripts/package-lock.json"
-          registry-url: "https://npm.pkg.github.com"
 
       - name: Setup python v3.11
         uses: actions/setup-python@8d9ed9ac5c53483de85588cdf95a591a75ab9f55 # v5.5.0
@@ -44,8 +38,6 @@ jobs:
 
       - name: Install dependencies
         run: npm ci
-        env:
-          NODE_AUTH_TOKEN: ${{ github.token }}
 
       - name: Run linting checks
         run: npm run lint
