+++ .github/workflows/typedoc-publish.yml
@@ -14,7 +14,6 @@ defaults:
 
 permissions:
   contents: read
-  packages: read
   pages: write
   id-token: write
 
@@ -29,14 +28,9 @@ jobs:
         uses: actions/setup-node@cdca7365b2dadb8aad0a33bc7601856ffabcc48e # v4.3.0
         with:
           node-version: 20
-          cache: npm
-          cache-dependency-path: "deploy/scripts/package-lock.json"
-          registry-url: "https://npm.pkg.github.com"
 
       - name: Install dependencies
         run: npm ci
-        env:
-          NODE_AUTH_TOKEN: ${{ github.token }}
 
       - name: Create TypeDoc documentation
         run: npx typedoc ./src/common/utils/**/*.ts
