+++ deploy/scripts/src/spot/requestGenerator/spotReqGen.ts
@@ -1,0 +1,32 @@
+import { type SpotRequest } from '../../spot/requestGenerator/spotReqFormat'
+import { uuidv4 } from '../../common/utils/jslib/index'
+
+export function generateRequest(currTime: string): SpotRequest {
+  const audClientID = uuidv4()
+  const sampleSpotRequest: SpotRequest = {
+    in_claims: {
+      'https://vocab.account.gov.uk/v1/credentialJWT': [
+        'eyJhbGciOiJFUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJ1cm46ZmRjOmdvdi51azoyMDIyOnMzQkkwQjh3NHNvUEdnQm5aX3ktVXVyMWJKNWFYRUN2aUNram1XZThNR2ciLCJuYmYiOjE2OTEzMTcyOTAsImlzcyI6InBhc3Nwb3J0Y3JpIiwidm90IjoiUDIiLCJleHAiOjE2OTQwMjA4OTAsImlhdCI6MTY5MTM0MjQ5MCwidmMiOnsiZXZpZGVuY2UiOlt7ImFjdGl2aXR5SGlzdG9yeVNjb3JlIjowLCJ2YWxpZGl0eVNjb3JlIjoyLCJ2ZXJpZmljYXRpb25TY29yZSI6MCwic3RyZW5ndGhTY29yZSI6NCwidHhuIjoidHhuIiwiaWRlbnRpdHlGcmF1ZFNjb3JlIjowLCJ0eXBlIjoiSWRlbnRpdHlDaGVjayJ9XSwiY3JlZGVudGlhbFN1YmplY3QiOnsibmFtZSI6W3sibmFtZVBhcnRzIjpbeyJ0eXBlIjoiR2l2ZW5OYW1lIiwidmFsdWUiOiJKYW5lIn0seyJ0eXBlIjoiRmFtaWx5TmFtZSIsInZhbHVlIjoiV3JpZ2h0In1dLCJ2YWxpZEZyb20iOiIyMDE5LTA0LTAxIn0seyJ2YWxpZFVudGlsIjoiMjAyOS0wNC0wMSIsIm5hbWVQYXJ0cyI6W3sidHlwZSI6IkdpdmVuTmFtZSIsInZhbHVlIjoiSmFuZSJ9LHsidHlwZSI6IkZhbWlseU5hbWUiLCJ2YWx1ZSI6IldyaWdodCJ9XX1dLCJiaXJ0aERhdGUiOlt7InZhbHVlIjoiMTk4OS0wNy0wNiJ9XX0sInR5cGUiOlsiVmVyaWZpYWJsZUNyZWRlbnRpYWwiLCJJZGVudGl0eUNoZWNrQ3JlZGVudGlhbCJdLCJAY29udGV4dCI6WyJodHRwczovL3d3dy53My5vcmcvMjAxOC9jcmVkZW50aWFscy92MSIsImh0dHBzOi8vdm9jYWIubG9uZG9uLmNsb3VkYXBwcy5kaWdpdGFsL2NvbnRleHRzL2lkZW50aXR5LXYxLmpzb25sZCJdfSwidnRtIjoiaHR0cHM6Ly9vaWRjLmFjY291bnQuZ292LnVrL3RydXN0bWFyayJ9.MEUCIEo_YKna0uk1lXW5KbUo63_-uQOXf3ibpL_YKein3mckAiEA2OamCs9WOcR5dKO8DiW9UCDqfqXMXxa9ZR-0VBoVD4o', // pragma: allowlist secret
+        'eyJhbGciOiJFUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJ1cm46ZmRjOmdvdi51azoyMDIyOkNZN1E2TDd4MEFlQWNmYUdlazY2R2FKeVhhM2R4RVhBaDRxWWtETUVzMUUiLCJuYmYiOjE2OTEzMTcyOTAsImlzcyI6ImZyYXVkY3JpIiwidm90IjoiUDIiLCJleHAiOjE2OTQwMjA4OTAsImlhdCI6MTY5MTM0MjQ5MCwidmMiOnsiZXZpZGVuY2UiOlt7ImFjdGl2aXR5SGlzdG9yeVNjb3JlIjowLCJ2YWxpZGl0eVNjb3JlIjowLCJ2ZXJpZmljYXRpb25TY29yZSI6MCwic3RyZW5ndGhTY29yZSI6MCwidHhuIjoidHhuIiwiaWRlbnRpdHlGcmF1ZFNjb3JlIjoxLCJ0eXBlIjoiSWRlbnRpdHlDaGVjayJ9XSwiY3JlZGVudGlhbFN1YmplY3QiOnsibmFtZSI6W3sibmFtZVBhcnRzIjpbeyJ0eXBlIjoiR2l2ZW5OYW1lIiwidmFsdWUiOiJKYW5lIn0seyJ0eXBlIjoiRmFtaWx5TmFtZSIsInZhbHVlIjoiV3JpZ2h0In1dLCJ2YWxpZEZyb20iOiIyMDE5LTA0LTAxIn0seyJ2YWxpZFVudGlsIjoiMjAyOS0wNC0wMSIsIm5hbWVQYXJ0cyI6W3sidHlwZSI6IkdpdmVuTmFtZSIsInZhbHVlIjoiSmFuZSJ9LHsidHlwZSI6IkZhbWlseU5hbWUiLCJ2YWx1ZSI6IldyaWdodCJ9XX1dLCJiaXJ0aERhdGUiOlt7InZhbHVlIjoiMTk4OS0wNy0wNiJ9XX0sInR5cGUiOlsiVmVyaWZpYWJsZUNyZWRlbnRpYWwiLCJJZGVudGl0eUNoZWNrQ3JlZGVudGlhbCJdLCJAY29udGV4dCI6WyJodHRwczovL3d3dy53My5vcmcvMjAxOC9jcmVkZW50aWFscy92MSIsImh0dHBzOi8vdm9jYWIubG9uZG9uLmNsb3VkYXBwcy5kaWdpdGFsL2NvbnRleHRzL2lkZW50aXR5LXYxLmpzb25sZCJdfSwidnRtIjoiaHR0cHM6Ly9vaWRjLmFjY291bnQuZ292LnVrL3RydXN0bWFyayJ9.MEQCIG9NNOV1KxVDyklTPjfuJL0ey01wsrPxe5m2VUPtL880AiAKG_n-Zalw7J-QX1G0FTBbMXur-QJii9vY8FwLzpHvQw', // pragma: allowlist secret
+        'eyJhbGciOiJFUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJ1cm46ZmRjOmdvdi51azoyMDIyOlBWUUk3Z0VmR1o2a080X0Z4eDZjaHF0SGNobzVzcjgtazA0MUdxci1YbzQiLCJuYmYiOjE2OTEzMTcyOTAsImlzcyI6InZlcmlmaWNhdGlvbmNyaSIsInZvdCI6IlAyIiwiZXhwIjoxNjk0MDIwODkwLCJpYXQiOjE2OTEzNDI0OTAsInZjIjp7ImV2aWRlbmNlIjpbeyJhY3Rpdml0eUhpc3RvcnlTY29yZSI6MCwidmFsaWRpdHlTY29yZSI6MCwidmVyaWZpY2F0aW9uU2NvcmUiOjIsInN0cmVuZ3RoU2NvcmUiOjAsInR4biI6InR4biIsImlkZW50aXR5RnJhdWRTY29yZSI6MCwidHlwZSI6IklkZW50aXR5Q2hlY2sifV0sImNyZWRlbnRpYWxTdWJqZWN0Ijp7Im5hbWUiOlt7Im5hbWVQYXJ0cyI6W3sidHlwZSI6IkdpdmVuTmFtZSIsInZhbHVlIjoiSmFuZSJ9LHsidHlwZSI6IkZhbWlseU5hbWUiLCJ2YWx1ZSI6IldyaWdodCJ9XSwidmFsaWRGcm9tIjoiMjAxOS0wNC0wMSJ9LHsidmFsaWRVbnRpbCI6IjIwMjktMDQtMDEiLCJuYW1lUGFydHMiOlt7InR5cGUiOiJHaXZlbk5hbWUiLCJ2YWx1ZSI6IkphbmUifSx7InR5cGUiOiJGYW1pbHlOYW1lIiwidmFsdWUiOiJXcmlnaHQifV19XSwiYmlydGhEYXRlIjpbeyJ2YWx1ZSI6IjE5ODktMDctMDYifV19LCJ0eXBlIjpbIlZlcmlmaWFibGVDcmVkZW50aWFsIiwiSWRlbnRpdHlDaGVja0NyZWRlbnRpYWwiXSwiQGNvbnRleHQiOlsiaHR0cHM6Ly93d3cudzMub3JnLzIwMTgvY3JlZGVudGlhbHMvdjEiLCJodHRwczovL3ZvY2FiLmxvbmRvbi5jbG91ZGFwcHMuZGlnaXRhbC9jb250ZXh0cy9pZGVudGl0eS12MS5qc29ubGQiXX0sInZ0bSI6Imh0dHBzOi8vb2lkYy5hY2NvdW50Lmdvdi51ay90cnVzdG1hcmsifQ.MEUCIH8jE2ayLS5c_NBzd8BPUodg6GHeY6xbDgcHf71CcbhUAiEAhm3COkSDQpCrsUNUtqjQv2E0nr4A5m5Y714IJKEJvdE', // pragma: allowlist secret
+        'eyJhbGciOiJFUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJ1cm46ZmRjOmdvdi51azoyMDIyOlY3RjMtX29uTC03YWx0VF9IejlLdjJlcDlxT2hFb0Y0WWcxelhYS3BERVEiLCJpc3MiOiJzZWN1cml0eWNyaSIsIm5iZiI6MTY5MTMxNzI5MCwiZXhwIjoxNjk0MDIwODkwLCJpYXQiOjE2OTEzNDI0OTAsInZvdCI6IlAyIiwidnRtIjoiaHR0cHM6Ly9vaWRjLmFjY291bnQuZ292LnVrL3RydXN0bWFyayIsInZjIjp7InR5cGUiOlsiVmVyaWZpYWJsZUNyZWRlbnRpYWwiLCJTZWN1cml0eUNoZWNrQ3JlZGVudGlhbCJdLCJjcmVkZW50aWFsU3ViamVjdCI6eyJuYW1lIjpbeyJuYW1lUGFydHMiOlt7InZhbHVlIjoiSmFuZSIsInR5cGUiOiJHaXZlbk5hbWUifSx7InZhbHVlIjoiV3JpZ2h0IiwidHlwZSI6IkZhbWlseU5hbWUifV0sInZhbGlkRnJvbSI6IjIwMTktMDQtMDEifSx7Im5hbWVQYXJ0cyI6W3sidmFsdWUiOiJKYW5lIiwidHlwZSI6IkdpdmVuTmFtZSJ9LHsidmFsdWUiOiJKZW5raW5zIiwidHlwZSI6IkZhbWlseU5hbWUifV0sInZhbGlkVW50aWwiOiIyMDI5LTA0LTAxIn1dLCJiaXJ0aERhdGUiOlt7InZhbHVlIjoiMTk4OS0wNy0wNiJ9XX0sImV2aWRlbmNlIjpbeyJ0eXBlIjoiU2VjdXJpdHlDaGVjayIsInR4biI6InR4biIsImNvbnRyYUluZGljYXRvcnMiOlt7ImNvZGUiOiJBMDEiLCJpc3N1YW5jZURhdGUiOiIyMDIyLTA5LTIwVDE1OjU0OjUwLjAwMFoiLCJkb2N1bWVudCI6InBhc3Nwb3J0L0dCUi84MjQxNTkxMjEiLCJtaXRpZ2F0aW9uIjp7Im1pdGlnYXRpbmdDcmVkZW50aWFsIjp7ImlzcyI6ImNyaS11cmkiLCJpYXQiOiIyMDIyLTA5LTIwVDE1OjU0OjUwLjAwMFoifX19XX1dfX0.MEQCICN2iK7o5jwlbw7dlcN4sFZdQuE1T5XLNqwQimucGX8GAiAnlh7hfN8XipJbQRgQs5VfqVEU5HTLHlJPbt2kLnuOsA' // pragma: allowlist secret
+      ],
+      vot: 'P2',
+      vtm: 'https://local.vtm'
+    },
+    in_local_account_id: 'a-simple-local-account-id',
+    in_rp_sector_id: 'a.simple.sector.id',
+    in_salt: 'YS1zaW1wbGUtc2FsdA==',
+    out_audience: audClientID,
+    log_ids: {
+      client_id: audClientID,
+      persistent_session_id: uuidv4(),
+      request_id: uuidv4() + '_' + currTime,
+      session_id: uuidv4(),
+      client_session_id: uuidv4()
+    },
+    out_sub: 'urn:fdc:gov.uk:2022:JG0RJI1pYbnanbvPs-j4j5-a-PFcmhry9Qu9NCEp5d4'
+  }
+
+  return sampleSpotRequest
+}
